+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This is the MuddyBot project changelog! Also the way I version is confusing so its not your fault its mine.
 
+## [1.0.4+snapshot25w25d] - 2025-06-17
+
+### Added
+
+- Now removes emojis from chat log file
+
+### Removed
+
+- Removed the !unlurk command
+- Removed the welcome message for users
+
 ## [1.0.3+snapshot25w25c] - 2025-06-16
 
 ### FIXED
@@ -170,4 +181,5 @@ _Initial release_
 [1.0.0]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.0
 [1.0.1+snapshot25w25a]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.1+snapshot25w25a
 [1.0.2+snapshot25w25b]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.2+snapshot25w25b
-[1.0.3+snapshot25w25c]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.2+snapshot25w25c
+[1.0.3+snapshot25w25c]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.3+snapshot25w25c
+[1.0.4+snapshot25w25d]: https://github.com/TheCrunching/python-twitch-bot/releases/tag/v1.0.4+snapshot25w25d